Revert "Question 5 to 10. (Fonctions intermediaires basiques)"
This reverts commit 693fee773fb85e3255d898eeb473137e62728d48.
