+++ build.xml
@@ -7,8 +7,8 @@
 <!-- the Compile on Save feature is turned off for the project. -->
 <!-- You can turn off the Compile on Save (or Deploy on Save) setting -->
 <!-- in the project's Project Properties dialog box.-->
-<project name="TP1Hopital" default="default" basedir=".">
-    <description>Builds, tests, and runs the project TP1Hopital.</description>
+<project name="2017-2018-poo4-seance1" default="default" basedir=".">
+    <description>Builds, tests, and runs the project 2017-2018-poo4-seance1.</description>
     <import file="nbproject/build-impl.xml"/>
     <!--
 
@@ -58,7 +58,7 @@
 
     An example of overriding the target for project execution could look like this:
 
-        <target name="run" depends="TP1Hopital-impl.jar">
+        <target name="run" depends="2017-2018-poo4-seance1-impl.jar">
             <exec dir="bin" executable="launcher.exe">
                 <arg file="${dist.jar}"/>
             </exec>
