--- createDDL.jdbc
@@ -1,6 +1,0 @@
-CREATE TABLE SERVICE (SERVNO INTEGER GENERATED BY DEFAULT AS IDENTITY NOT NULL, LOCALISATION VARCHAR(25) NOT NULL, NOM VARCHAR(255) UNIQUE, ID_MEDECIN_DIRIGE INTEGER, PRIMARY KEY (SERVNO))
-CREATE TABLE MEDECIN (MEDNO INTEGER GENERATED BY DEFAULT AS IDENTITY NOT NULL, NOM VARCHAR(255) NOT NULL, PRENOM VARCHAR(255) NOT NULL, SALAIRE FLOAT, ID_MEDECIN_CHEF INTEGER, ID_SERVICE INTEGER, PRIMARY KEY (MEDNO))
-ALTER TABLE MEDECIN ADD CONSTRAINT UNQ_MEDECIN_0 UNIQUE (PRENOM, NOM)
-ALTER TABLE SERVICE ADD CONSTRAINT SRVCEDMDECINDIRIGE FOREIGN KEY (ID_MEDECIN_DIRIGE) REFERENCES MEDECIN (MEDNO)
-ALTER TABLE MEDECIN ADD CONSTRAINT MEDECIN_ID_SERVICE FOREIGN KEY (ID_SERVICE) REFERENCES SERVICE (SERVNO)
-ALTER TABLE MEDECIN ADD CONSTRAINT MDECINDMEDECINCHEF FOREIGN KEY (ID_MEDECIN_CHEF) REFERENCES MEDECIN (MEDNO)
