--- dropDDL.jdbc
@@ -1,6 +1,0 @@
-ALTER TABLE SERVICE DROP CONSTRAINT SRVCEDMDECINDIRIGE
-ALTER TABLE MEDECIN DROP CONSTRAINT MEDECIN_ID_SERVICE
-ALTER TABLE MEDECIN DROP CONSTRAINT MDECINDMEDECINCHEF
-ALTER TABLE MEDECIN DROP CONSTRAINT UNQ_MEDECIN_0
-DROP TABLE SERVICE
-DROP TABLE MEDECIN
