+++ nbproject/project.properties
@@ -4,8 +4,8 @@ annotation.processing.processor.options=-Aeclipselink.canonicalmodel.use_static_
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
-application.title=TP1Hopital
-application.vendor=Ledieu
+application.title=2017-2018-poo4-seance1
+application.vendor=user
 build.classes.dir=${build.dir}/classes
 build.classes.excludes=**/*.java,**/*.form
 # This directory is removed when the project is cleaned:
@@ -26,7 +26,7 @@ debug.test.classpath=\
 dist.archive.excludes=
 # This directory is removed when the project is cleaned:
 dist.dir=dist
-dist.jar=${dist.dir}/TP1Hopital.jar
+dist.jar=${dist.dir}/2017-2018-poo4-seance1.jar
 dist.javadoc.dir=${dist.dir}/javadoc
 endorsed.classpath=
 excludes=
@@ -60,11 +60,12 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=hopital.Hopital
+main.class=centralellile._2017_2018.poo4.seance1.Main
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
 platform.active=default_platform
+project.license=mit
 run.classpath=\
     ${javac.classpath}:\
     ${build.classes.dir}
